Cleaned up code and added helper functions
